* 'develop' of github.com:ampproject/amp-wp:
  Re-switch php-compatibility from wimg to phpcompatibility; pin at v9.1.1
  Remove unnecessary CommentedOutCode ignore
  Revert "Update WPCS and PHPCompatibility to latest versions and rely on lock file for concrete versions."
  Fix all outstanding coding standards violations, via PHPCBF, adjustments to phpcs.xml and manual modifications.
  Remove outdated WPCS annotations that are no longer preferred.
  Update WPCS and PHPCompatibility to latest versions and rely on lock file for concrete versions.

+++ includes/sanitizers/class-amp-tag-and-attribute-sanitizer.php
@@ -297,18 +297,23 @@ class AMP_Tag_And_Attribute_Sanitizer extends AMP_Base_Sanitizer {
 				AMP_Rule_Spec::MANDATORY => true,
 			);
 
-			$versions = array_unique( array_merge(
-				isset( $extension_spec['allowed_versions'] ) ? $extension_spec['allowed_versions'] : array(),
-				isset( $extension_spec['version'] ) ? $extension_spec['version'] : array()
-			) );
+			$versions = array_unique(
+				array_merge(
+					isset( $extension_spec['allowed_versions'] ) ? $extension_spec['allowed_versions'] : array(),
+					isset( $extension_spec['version'] ) ? $extension_spec['version'] : array()
+				)
+			);
 
 			$rule_spec[ AMP_Rule_Spec::ATTR_SPEC_LIST ]['src'] = array(
-				AMP_Rule_Spec::VALUE_REGEX => implode( '', array(
-					'^',
-					preg_quote( 'https://cdn.ampproject.org/v0/' . $extension_spec['name'] . '-' ),
-					'(' . implode( '|', $versions ) . ')',
-					'\.js$',
-				) ),
+				AMP_Rule_Spec::VALUE_REGEX => implode(
+					'',
+					array(
+						'^',
+						preg_quote( 'https://cdn.ampproject.org/v0/' . $extension_spec['name'] . '-' ), // phpcs:ignore WordPress.PHP.PregQuoteDelimiter.Missing
+						'(' . implode( '|', $versions ) . ')',
+						'\.js$',
+					)
+				),
 			);
 		}
 
@@ -1373,13 +1378,15 @@ class AMP_Tag_And_Attribute_Sanitizer extends AMP_Base_Sanitizer {
 		 * unless the commas are URL-encoded.
 		 */
 		if ( 'srcset' === $attribute_node->nodeName || 'srcset' === $spec_attr_name ) {
-			return array_filter( array_map(
-				function ( $srcset_part ) {
-					// Remove descriptors for width and pixel density.
-					return preg_replace( '/\s.*$/', '', trim( $srcset_part ) );
-				},
-				preg_split( '/\s*,\s*/', $attribute_node->nodeValue )
-			) );
+			return array_filter(
+				array_map(
+					function ( $srcset_part ) {
+						// Remove descriptors for width and pixel density.
+						return preg_replace( '/\s.*$/', '', trim( $srcset_part ) );
+					},
+					preg_split( '/\s*,\s*/', $attribute_node->nodeValue )
+				)
+			);
 		} else {
 			return array( $attribute_node->nodeValue );
 		}
@@ -1579,7 +1586,7 @@ class AMP_Tag_And_Attribute_Sanitizer extends AMP_Base_Sanitizer {
 					$required_value = $property_spec['value'];
 				} elseif ( isset( $property_spec['value_double'] ) ) {
 					$required_value = $property_spec['value_double'];
-					$prop_value     = (double) $prop_value;
+					$prop_value     = (float) $prop_value;
 				}
 				if ( isset( $required_value ) && $prop_value !== $required_value ) {
 					return AMP_Rule_Spec::FAIL;
